Added get all API staff member

+++ FinalWebAPI/Models/StaffMember.cs
@@ -4,7 +4,7 @@
     {
         public int? Id { get; set; }
         public string? MemberName { get; set; }
-        public int? Mobile { get; set; }
+        public long? Mobile { get; set; }
         public string? StaffAddress { get; set; }
         public string? MemberServices { get; set; }
         public decimal? Salary { get; set; }
